Update bar chart exmaple
